Hand-roll big-endian deref implementation
Use the count trailing zeros method on big-endian architectures instead
of byte-swapping the integer to little-endian. This passes all tests
when ran with `miri test --target mips64-unknown-linux-gnuabi64`.

I am unable to check if this improves performance because I do not own
a big-endian machine to benchmark with. Oddly on (mips)[https://rust.godbolt.org/z/9bE3naex9]
the trailing_zeros method produces more assembly because mips does not have a count trailing zeros instruction
but it this seems like a implementation detail that can be dealt with by the codegen backend.